\name{print.summary.parfrailty}

\alias{print.summary.parfrailty}

\title{
Prints summary of parfrailty fit
}

\description{
This is a \code{print} method for class \code{"summary.parfrailty"}.
}

\usage{
\method{print}{summary.parfrailty}(x, digits = max(3L, getOption("digits") - 3L),
                             ...)
}

\arguments{
  \item{x}{
an object of class \code{"summary.parfrailty"}.
}
  \item{digits}{
the number of significant digits to use when printing.
}
  \item{\dots}{
not used.
}
}

\author{
Arvid Sjolander and Elisabeth Dahlqwist
}

\seealso{
\code{\link{parfrailty}}
}

\examples{

##See documentation for frailty

}
