\name{print.summary.stdCoxph}

\alias{print.summary.stdCoxph}

\title{
Prints summary of Cox regression standardization fit
}

\description{
This is a \code{print} method for class \code{"summary.stdCoxph"}.
}

\usage{
\method{print}{summary.stdCoxph}(x, ...)
}

\arguments{
  \item{x}{
an object of class \code{"summary.stdCoxph"}.
}
  \item{\dots}{
not used.
}
}

\author{
Arvid Sjolander
}

\seealso{
\code{\link{stdCoxph}}
}

\examples{

##See documentation for stdCoxph

}
