\name{print.summary.stdGee}

\alias{print.summary.stdGee}

\title{
Prints summary of GEE regression standardization fit
}

\description{
This is a \code{print} method for class \code{"summary.stdGee"}.
}

\usage{
\method{print}{summary.stdGee}(x, ...)
}

\arguments{
  \item{x}{
an object of class \code{"summary.stdGee"}.
}
  \item{\dots}{
not used.
}
}

\author{
Arvid Sjolander
}

\seealso{
\code{\link{stdGee}}
}

\examples{

##See documentation for stdGee

}
