\name{print.summary.stdGlm}

\alias{print.summary.stdGlm}

\title{
Prints summary of GLM regression standardization fit
}

\description{
This is a \code{print} method for class \code{"summary.stdGlm"}.
}

\usage{
\method{print}{summary.stdGlm}(x, ...)
}

\arguments{
  \item{x}{
an object of class \code{"summary.stdGlm"}.
}
  \item{\dots}{
not used.
}
}

\author{
Arvid Sjolander
}

\seealso{
\code{\link{stdGlm}}
}

\examples{

##See documentation for stdGlm

}
