\name{print.summary.stdParfrailty}

\alias{print.summary.stdParfrailty}

\title{
Prints summary of Frailty standardization fit
}

\description{
This is a \code{print} method for class \code{"summary.stdParfrailty"}.
}

\usage{
\method{print}{summary.stdParfrailty}(x, ...)
}

\arguments{
  \item{x}{
an object of class \code{"summary.stdParfrailty"}.
}
  \item{\dots}{
not used.
}
}

\author{
Arvid Sjolander 
}

\seealso{
\code{\link{stdParfrailty}}
}

\examples{

##See documentation for stdParfrailty

}
