\name{stdReg-package}
\alias{stdReg-package}
\alias{stdReg}
\docType{package}
\title{
\packageTitle{stdReg}
}
\description{
\packageDescription{stdReg}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{stdReg}
\packageIndices{stdReg}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{stdReg}

Maintainer: \packageMaintainer{stdReg}
}
\references{

Rothman K.J., Greenland S., Lash T.L. (2008). \emph{Modern Epidemiology}, 3rd edition. 
Lippincott, Williams \& Wilkins.

Makuch R.W. (1982). Adjusted survival curve estimation using covariates.
\emph{Journal of Chronic Diseases} \bold{35}, 437-443.

Chang I.M., Gelman G., Pagano M. (1982). Corrected group prognostic curves
and summary statistics. \emph{Journal of Chronic Diseases} \bold{35}, 669-674.

Gail M.H. and Byar D.P. (1986). Variance calculations for direct adjusted survival
curves, with applications to testing for no treatement effect. \emph{Biometrical Journal}
\bold{28}(5), 587-599. 

}
\keyword{ package }

